Ajustes de uso nas Utils de verificação, aplicando interfaces nos mappers e fazendo o service lidar com entities

+++ src/main/java/repository/AnuncioRepository.java
@@ -3,8 +3,10 @@ package repository;
 import dao.AnuncioDAO;
 import dao.DeslikeDAO;
 import dao.LikeDAO;
+import dto.DTOUtils;
 import model.AnuncioEntity;
 import repository.contracts.AdRepository;
+import repository.exceptions.NullEntityForRepository;
 
 import java.util.ArrayList;
 
@@ -13,7 +15,11 @@ public class AnuncioRepository implements AdRepository {
     private final LikeDAO likeDAO;
     private final DeslikeDAO deslikeDAO;
 
-    public AnuncioRepository(AnuncioDAO anuncioDAO, LikeDAO likeDAO, DeslikeDAO deslikeDAO) {
+    public AnuncioRepository(
+            AnuncioDAO anuncioDAO,
+            LikeDAO likeDAO,
+            DeslikeDAO deslikeDAO
+    ) {
         this.anuncioDAO = anuncioDAO;
         this.likeDAO = likeDAO;
         this.deslikeDAO = deslikeDAO;
@@ -27,29 +33,48 @@ public class AnuncioRepository implements AdRepository {
 
     @Override
     public void removerAnuncio(AnuncioEntity anuncio) {
+        if(isNullEntity(anuncio))
+            throw new NullEntityForRepository();
+
         anuncioDAO.deleteAnuncio(anuncio.getIdVendedor(), anuncio.getTitulo());
     }
 
     @Override
     public AnuncioEntity lerAnuncioPeloId(int id) {
-        AnuncioEntity anuncio = anuncioDAO.readAnuncioById(id);
-        anuncio.setLikes(likeDAO.readLikesOfAd(anuncio.getId()));
-        anuncio.setDeslikes(deslikeDAO.readDeslikesOfAd(anuncio.getId()));
+        AnuncioEntity anuncioByDAO = anuncioDAO.readAnuncioById(id);
 
-        return anuncio;
+        if(isNullEntity(anuncioByDAO))
+            return null;
+
+        anuncioByDAO.setLikes(likeDAO.readLikesOfAd(id));
+        anuncioByDAO.setDeslikes(deslikeDAO.readDeslikesOfAd(id));
+
+        return returnAnuncioWithCheck(anuncioByDAO);
     }
 
     @Override
     public AnuncioEntity lerAnuncioPeloNome(String nome) {
-        AnuncioEntity anuncio = anuncioDAO.readAnuncioByName(nome);
-        anuncio.setLikes(likeDAO.readLikesOfAd(anuncio.getId()));
-        anuncio.setDeslikes(deslikeDAO.readDeslikesOfAd(anuncio.getId()));
+        AnuncioEntity anuncioByDAO = anuncioDAO.readAnuncioByName(nome.toUpperCase());
 
-        return anuncio;
+        if(isNullEntity(anuncioByDAO))
+            return null;
+
+        anuncioByDAO.setLikes(likeDAO.readLikesOfAd(anuncioByDAO.getId()));
+        anuncioByDAO.setDeslikes(deslikeDAO.readDeslikesOfAd(anuncioByDAO.getId()));
+
+        return returnAnuncioWithCheck(anuncioByDAO);
     }
 
     @Override
     public ArrayList<AnuncioEntity> lerTodosOsAnuncios() {
         return anuncioDAO.readAllFromAnuncios();
+    }
+
+    private boolean isNullEntity(AnuncioEntity anuncio){
+        return anuncio == null;
+    }
+
+    private AnuncioEntity returnAnuncioWithCheck(AnuncioEntity anuncio){
+        return  (isNullEntity(anuncio)) ? null : anuncio;
     }
 }
